fix: corrigiendo tipo de dato en SP

+++ src/payments/registerPayment.ts
@@ -1,4 +1,4 @@
-import { Int, Date } from 'mssql';
+import { Int, Date, Float } from 'mssql';
 import { SPAltaPago } from './types/SPAltaPago';
 import { statusResponse } from './types/pagos';
 import { DbConnector } from "../helpers/dbConnector"
@@ -16,7 +16,7 @@ export const registerPayment = async (
       .input('NUMERO_SEMANA', Int, spaAltaPago.NUMERO_SEMANA)
       .input('ID_CLIENTE', Int, spaAltaPago.ID_CLIENTE)
       .input('ID_USUARIO', Int, spaAltaPago.ID_USUARIO)
-      .input('CANTIDAD_PAGADA', Int, spaAltaPago.CANTIDAD_PAGADA)
+      .input('CANTIDAD_PAGADA', Float, spaAltaPago.CANTIDAD_PAGADA)
       .input('FECHA_ACTUAL', Date, spaAltaPago.FECHA_ACTUAL)
       .input('ID_COBRADOR', Int, spaAltaPago.ID_COBRADOR)
       .execute('SP_ALTA_PAGO');
